+++ tests/sentry/api/test_event_search.py
@@ -10,12 +10,10 @@ from django.utils import timezone
 from sentry.api.event_search import (
     AggregateFilter,
     AggregateKey,
-    ParenExpression,
     SearchConfig,
     SearchFilter,
     SearchKey,
     SearchValue,
-    cleanup_search_query,
     parse_search_query,
 )
 from sentry.constants import MODULE_ROOT
@@ -707,45 +705,3 @@ def test_search_value_to_query_string(value, expected_query_string):
     actual = search_value.to_query_string()
 
     assert actual == expected_query_string
-
-
-@pytest.mark.parametrize(
-    "dirty, clean",
-    [
-        ("release:initial OR os.name:android", "release:initial OR os.name:android"),
-        ("OR AND OR release:initial OR os.name:android", "release:initial OR os.name:android"),
-        ("release:initial OR os.name:android AND OR AND ", "release:initial OR os.name:android"),
-        (
-            "release:initial AND (AND OR) (OR )os.name:android ",
-            "release:initial AND os.name:android",
-        ),
-        (
-            " AND ((AND OR (OR ))) release:initial (((AND OR  (AND)))) AND os.name:android  (AND OR) ",
-            "release:initial AND os.name:android",
-        ),
-        (" (AND) And (And) Or release:initial or (and) or", "release:initial"),
-    ],
-)
-def test_cleanup_query(dirty, clean):
-    dirty_tokens = parse_search_query(dirty)
-    clean_tokens = parse_search_query(clean)
-    actual_clean = cleanup_search_query(dirty_tokens)
-
-    assert actual_clean == clean_tokens
-
-
-def test_cleanup_query_with_empty_parens():
-    """
-    Separate test with empty parens because we can't parse a string with empty parens correctly
-    """
-    paren = ParenExpression
-    dirty_tokens = (
-        [paren([paren(["AND", "OR", paren([])])])]
-        + parse_search_query("release:initial AND (AND OR) (OR)")  # ((AND OR (OR ())))
-        + [paren([])]
-        + parse_search_query("os.name:android")  # ()
-        + [paren([paren([paren(["AND", "OR", paren([])])])])]  # ((()))
-    )
-    clean_tokens = parse_search_query("release:initial AND os.name:android")
-    actual_clean = cleanup_search_query(dirty_tokens)
-    assert actual_clean == clean_tokens
